Make float

+++ gls/gls.py
@@ -224,7 +224,7 @@ def main():
     parser = argparse.ArgumentParser(description="Run gate-level simulation for CGRA: verification and power estimation")
     parser.add_argument('app', help="Name of CGRA application to test", type=str)
     parser.add_argument('-acn', '--active_cycle_number', help="Number of cycles after which to measure toggling activity, typically after CGRA configuration", required=False, type=int, default=None)
-    parser.add_argument('-acf', '--app_clk_factor', help="How much slower should timing be for this app", required=False, type=int, default=None)
+    parser.add_argument('-acf', '--app_clk_factor', help="How much slower should timing be for this app", required=False, type=float, default=None)
     parser.add_argument('--merge-saif-only', help="Only do the merge SAIF step", required=False, action='store_true')
     args = parser.parse_args()
